// src/app/interfaces/productos.interface.ts
// Generated by https://quicktype.io

export interface PageProductosSimple {
  total:      number;
  totalPages: number;
  page:       number;
  pageSize:   number;
  productos:  Producto[];
}

export interface Producto {
  id:             number;
  empresaId:      number;
  categoriaId:    number;
  subCategoriaId: number;
  marcaId:        number;
  nombre:         string;
  descripcion:    string;
  activo:         boolean;
  marca:          Marca;
  categoria:      Categoria;
  subCategoria:   SubCategoria;
}

export interface SubCategoria {
  id:           string;
  empresaId:    number;
  descripcion:  string;
  activo:       boolean;
  categoriaId?: number;
}
export interface Categoria {
  id:           string;
  empresaId:    number;
  descripcion:  string;
  activo:       boolean;
  categoriaId?: number;
}
export interface Marca {
  id:           string;
  empresaId:    number;
  descripcion:  string;
  activo:       boolean;
  categoriaId?: number;
}